[#9] Remove unused classes and move Interfaces to directory and change namepsace
