+++ CodeGenEngine/CodeGenEngine/Include.cs
@@ -1,4 +1,6 @@
-﻿namespace CodeGenEngine
+﻿using CodeGenEngine.Interface;
+
+namespace CodeGenEngine
 {
     public class Include : IElement, IMapped
     {
