+++ CodeGenEngine/CodeGenEngine/Interface/IMapped.cs
@@ -1,4 +1,4 @@
-﻿namespace CodeGenEngine
+﻿namespace CodeGenEngine.Interface
 {
     public interface IMapped
     {
