+++ CodeGenEngine/CodeGenEngine/Property.cs
@@ -1,4 +1,5 @@
 ﻿using CodeGenEngine.Abstract;
+using CodeGenEngine.Interface;
 using System;
 using System.Collections.Generic;
 using System.Linq;
